+++ eNMS/static/js/workflow.js
@@ -696,7 +696,9 @@ Object.assign(action, {
   "Edit Workflow": () => showTypePanel("workflow", workflow.id),
   "Restart Workflow from Here": (service) =>
     showRestartWorkflowPanel(workflow, service),
-  "Workflow Results": () => showRuntimePanel("results", workflow),
+  "Workflow Results Tree": () => showRuntimePanel("results", workflow),
+  "Workflow Results Table": () =>
+    showRuntimePanel("results", workflow, null, "full_result"),
   "Workflow Logs": () => showRuntimePanel("logs", workflow),
   "Add to Workflow": addServicePanel,
   "Stop Workflow": () => stopWorkflow(),