+++ tests/test_admin.py
@@ -30,7 +30,7 @@ def test_user_management(user_client):
             "form_type": "user",
             "name": user,
             "email": f"{user}@test.com",
-            "group": "Admin",
+            "groups": ["Admin"],
             "password": user,
         }
         user_client.post("/update/user", data=dict_user)